refactor(xtask): move the BINDIR name set + Os enum into xtask-lib for cross-crate reuse
